+++ src/lecture1/constructor/ex3/TestMain.java
@@ -1,0 +1,16 @@
+package lecture1.constructor.ex3;
+
+import lecture1.constructor.ex3.vo.Animal;
+import lecture1.constructor.ex3.vo.Human;
+
+public class TestMain {
+
+    public static void main(String[] args){
+        Animal animal = new Animal();
+        animal.runGOGO();
+
+        Human huMan = new Human();
+        huMan.runGOGO();
+    }
+
+}
