+++ src/lecture1/constructor/ex4/vo/NCUserInfo.java
@@ -1,8 +1,6 @@
-package lecture1.constructorlecture.ex4.vo;
+package lecture1.constructor.ex4.vo;
 
 public class NCUserInfo extends UserInfo {
-
-    private boolean isNexonPrime = false;
 
     public NCUserInfo(String id, String pw, String email) {
         super(id, pw, email);
