--- src/lecture1/constructorlecture/ex1/TestMain.java
@@ -1,10 +1,0 @@
-package lecture1.constructorlecture.ex1;
-
-public class TestMain {
-
-    public static void main(String args[]) {
-        UserInfo userInfo = new UserInfo("중2병도사", "9292");
-        userInfo.printId();
-    }
-
-}
