--- src/lecture1/constructorlecture/ex3/ABCDEFGMain.java
@@ -1,16 +1,0 @@
-package lecture1.constructorlecture.ex3;
-
-import lecture1.constructorlecture.ex3.vo.Animal;
-import lecture1.constructorlecture.ex3.vo.Human;
-
-public class ABCDEFGMain {
-
-    public static void main(String args[]){
-        Animal animal = new Animal();
-        animal.runGOGO();
-
-        Human huMan = new Human();
-        huMan.runGOGO();
-    }
-
-}
